Fix Hibernate sequence updates
Updating sequences to be at the top of any serial id, because Hibernate does not
do that automatically at startup. It just queries the max value of each sequence
to get the next value. In addition, Hibernate does not generate a proper schema to
auto increment on each insertion.
XXX workaround to prevent unique key constraint violations during insertion: Should
be changed ASAP because we need to hard-code DB specific sequence updates here!

Please note, the list of updated sequences is not complete. We should add to it as soon
as we get another key constraint exception.

Fixes #6