dal: Unnesting JPAExceptions
We do not want to nest JPAExceptions, which should report issues to the API user.
That is, it should always return the very first error, that occurred and got enough
information to solve an API call issue. We also want to show stack traces for debug
purposes, if it is not an already described exception (not an JPAException).