writer: use a application managed entity manager
In the past we used @PersistenceContext here. However, if we inject this
EntityManager like that in environments where the concurrency is a
concern, we're not thread-safe. This caused some invalid pushRecord
calls, where strings would have been pushed to number measurement tables
which resulted in a database error. Hence, we better use a application
managed entity manager here.